Separating Rsync server and moving from cores -> instances
